Perform Structure cleanup
