+++ documentation/workflow.drawio
@@ -1,0 +1,1 @@
+<mxfile host="app.diagrams.net" modified="2020-10-08T12:22:50.349Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_6) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/85.0.4183.121 Safari/537.36" etag="AZdswP0WqX9f8YjrKnlT" version="13.7.8" type="device"><diagram id="C5RBs43oDa-KdzZeNtuy" name="Page-1">7VhZc9MwEP41Hp5gfBQnfcRp0wKhU0ihw6NiK7aI7DWynINfzyqWYzsOOYD0yDQv0a43K2m/b4/YcHrx/EqQNPoEAeWGbQZzw7kwbNs2XRe/lGZRaCyzaxeaULBA6yrFkP2ipaHW5iygWcNQAnDJ0qbShyShvmzoiBAwa5qNgTd3TUlIW4qhT3hbe88CGRXart2p9NeUhVG5s+WeF09iUhrrm2QRCWBWUzmXhtMTALJYxfMe5Sp6ZVzu3y/u+WDiXn34nP0kX72PdzffXhfO+of8ZHUFQRP5f11rLKeE5zpe+q5yUQZQQJ4EVDkxDceLZMxxaeHyB5VyoQEnuQRUgZARhJAQPgBItd0YEqnNLCXTJHingEV5xMGfFKo+41zvgZK276KUSQGTFXbKwQoIZczJiHKP+JNwedAecBD4KIGEKlcBkkHfpTrcZaX19oytxiCDXPh0i52jKU5ESLf502mlzlfjqUbuikJMpViggaCcSDZtkpnonAhXdhXuuNDQH0ADp0WDAYlTlfZAs+SVxNUMxGQLNxQus4hJOkzJMkAzrCdNvtR5gFf1Qk6yTKO4A+TDQJpSIel8a1jLp67ObV3dulqcVZXCKtM/qlWJM/NIQJy1gPiOxfMlJf8lJd09U7LsWDtzclE2x8J+7xTVnm6BJbJmAuNxhgdbp85qw79n09sWm26gRabdeDxruj0SmTqHccl68lxy/9AiXI5h8EYCV6FapTwPQyQLuksMp9+uXRHEozzb3S4aGCsG9UnMuIrXNeVTKplPNjQVwlmI+174CDcVm8mDW7IkRMmtpLslWbH+HrHZdJrNZiXXu425odt0j9VtOi1MbxG9Ajrb5Et8T6zjO/YaCOeP3fK7+xTpZ12CH7zjl9V0d8u39izTmj3mGwc/DQI9/SngfK+Z8mUMOA7D7FMbBMoT1vjk5XzUngRGuUiWgwDk8mUS2DYJ2Jua0INOAmUhrIH6haaEIZjm4BTngPV//mfOY88BZaVoQsBVTG2zyLATw2B9FnOs42GAYvWmtiiF1Qtv5/I3</diagram></mxfile>
