+++ lib/ui/components/icon_label.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../constants.dart';
+import 'package:ForkifyApp/utilities/constants.dart';
 import 'package:flutter_svg/svg.dart';
