+++ lib/ui/home/search_header.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import '../constants.dart';
+import 'package:ForkifyApp/utilities/constants.dart';
 
 class SearchHeader extends StatelessWidget {
   const SearchHeader({
